Move data mutating logic into the rendering pipeline for full text
