+++ .rubocop_todo.yml
@@ -765,7 +765,6 @@ RSpec/ContextMethod:
     - 'spec/models/concerns/eds_links_spec.rb'
     - 'spec/models/concerns/eds_subjects_spec.rb'
     - 'spec/models/marc_fields/linked_related_works_spec.rb'
-    - 'spec/presenters/presenter_research_starter_spec.rb'
     - 'spec/routing/article_routes_spec.rb'
     - 'spec/services/eds/search_service_spec.rb'
 
@@ -928,7 +927,6 @@ RSpec/MultipleMemoizedHelpers:
 # This cop supports safe autocorrection (--autocorrect).
 RSpec/MultipleSubjects:
   Exclude:
-    - 'spec/presenters/presenter_research_starter_spec.rb'
     - 'spec/services/eds/search_service_spec.rb'
 
 # Offense count: 284
@@ -1026,7 +1024,6 @@ RSpec/SubjectStub:
     - 'spec/presenters/article_fulltext_link_presenter_spec.rb'
     - 'spec/presenters/facet_options_presenter_spec.rb'
     - 'spec/presenters/presenter_format_spec.rb'
-    - 'spec/presenters/presenter_research_starter_spec.rb'
 
 # Offense count: 93
 # Configuration parameters: IgnoreNameless, IgnoreSymbolicNames.
@@ -1661,7 +1658,6 @@ Style/RedundantInterpolation:
     - 'app/helpers/article_helper.rb'
     - 'app/helpers/xml_api_helper.rb'
     - 'app/views/catalog/opensearch.xml.builder'
-    - 'spec/presenters/presenter_research_starter_spec.rb'
 
 # Offense count: 6
 # This cop supports safe autocorrection (--autocorrect).
