--- app/views/articles/_show_reference.html.erb
@@ -1,4 +1,0 @@
-<% doc_presenter = document_presenter(document) -%>
-<%# we need to transform the EDS HTML Full Text content as it's in pseudo-HTML markup %>
-<% doc_presenter.transform_research_starter_text %>
-<%= render "articles/show_default", document: document -%>
