+++ spec/helpers/article_helper_spec.rb
@@ -222,4 +222,70 @@ RSpec.describe ArticleHelper do
       end
     end
   end
+
+  describe '#sanitize_research_starter' do
+    subject(:result) { view.sanitize_research_starter(value: [value], document: document) }
+
+    let(:document) { SolrDocument.new('eds_database_id' => 'xyz') }
+
+    context '#transform_research_starter_text' do
+      let(:value) { 'abc' }
+
+      context 'removes anid and title' do
+        let(:value) { '<anid>abc</anid><title>123</title>' }
+
+        it do
+          expect(result).to eq ''
+        end
+      end
+
+      context 'rewrites EDS tags with regular HTML tags' do
+        let(:value) { '<bold/><emph/><ulist><item><subs/><sups/></item></ulist>' }
+
+        it do
+          expect(result).to eq "<b></b><i></i><ul><li>\n<sub></sub><sup></sup>\n</li></ul>"
+        end
+      end
+
+      context 'rewrites EDS eplinks' do
+        let(:value) { '<eplink linkkey="abc123">def456</eplink>' }
+        let(:document) { { 'eds_html_fulltext' => value, 'eds_database_id' => 'xyz' } }
+
+        it do
+          expect(result).to eq '<a href="/articles/xyz__abc123">def456</a>'
+        end
+      end
+
+      context 'converts images to figures with captions' do
+        let(:value) { '<img src="abc.jpg" title="def"/>' }
+
+        it do
+          html = Capybara.string(result)
+          expect(html).to have_css('.research-starter-figure')
+          expect(html).to have_css('img[src="abc.jpg"]')
+          expect(html).to have_css('span', text: 'def')
+        end
+      end
+
+      context 'removes any unknown elements' do
+        let(:value) { '<unknown/>' }
+
+        it do
+          expect(result).to eq ''
+        end
+      end
+
+      context 'preserves desired element' do
+        %w[p a b i ul li div span sub sup].each do |tag|
+          context tag do
+            let(:value) { "<#{tag}>abc</#{tag}>" }
+
+            it do
+              expect(result).to eq(value).and be_html_safe
+            end
+          end
+        end
+      end
+    end
+  end
 end
